更新

+++ app/src/main/java/com/example/administrator/moblieplayer/utli/Utli.java
@@ -12,7 +12,7 @@ public class Utli {
         Toast.makeText(context,message,Toast.LENGTH_SHORT).show();
 
     }
-    public static String formatter(Context context,long size){
+    public static String formatter(Context context, int size){
         String fileSize;
         fileSize = android.text.format.Formatter.formatFileSize(context,size);
         return fileSize;
